feat: 자동차 이름을 부여할 수 있도록 추가
