+++ src/test/java/carrase/CarTest.java
@@ -1,16 +1,20 @@
 package carrase;
 
-import static org.assertj.core.api.AssertionsForClassTypes.*;
+import static org.assertj.core.api.Assertions.*;
+
+import java.util.Arrays;
+import java.util.List;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 public class CarTest {
+	final String name = "TEST";
 
 	@Test
 	@DisplayName("자동차 전진")
 	void carMove() {
-		RacingCar car = new RacingCar();
+		RacingCar car = new RacingCar(name);
 		car.moveForward();
 
 		assertThat(car.getMileage()).isEqualTo(1);
@@ -19,7 +23,7 @@ public class CarTest {
 	@Test
 	@DisplayName("자동차 멈춤")
 	void carStop() {
-		RacingCar car = new RacingCar();
+		RacingCar car = new RacingCar(name);
 		car.stop();
 
 		assertThat(car.getMileage()).isEqualTo(0);
@@ -29,7 +33,7 @@ public class CarTest {
 	@DisplayName("전진")
 	void move() {
 		CarBehavior carBehavior = CarBehavior.getCarBehavior(4);
-		RacingCar car = new RacingCar();
+		RacingCar car = new RacingCar(name);
 		carBehavior.apply(car);
 
 		assertThat(car.getMileage()).isEqualTo(1);
@@ -39,9 +43,17 @@ public class CarTest {
 	@DisplayName("멈춤")
 	void stop() {
 		CarBehavior carBehavior = CarBehavior.getCarBehavior(3);
-		RacingCar car = new RacingCar();
+		RacingCar car = new RacingCar(name);
 		carBehavior.apply(car);
 
 		assertThat(car.getMileage()).isEqualTo(0);
+	}
+
+	@Test
+	@DisplayName("자동차 이름은 5자를 초과할 수 없다")
+	void notExceedCarNameLengthFive() {
+		assertThatThrownBy(() -> {
+			new RacingCar("longName");
+		}).isInstanceOf(IllegalArgumentException.class);
 	}
 }
